fix bugs, log actions

+++ Filesystem/struct.h
@@ -1,10 +1,9 @@
 typedef struct inode
 {
-    char *name;
-    char *content;
-    struct inode *parent;
-    struct inode *childs;
-    unsigned int childs_c;
-
+	char* name;
+	char* content;
+	struct inode *parent;
+	struct inode *childs;
+	size_t childs_c;
 } inode_t;
